COL-168 - Moved cms-components under the cds project
