+++ projects/cds/src/cms-components/merchandising/carousel/index.ts
