+++ projects/cds/src/cms-components/merchandising/index.ts
