+++ projects/core/src/model/cms.model.ts
@@ -142,13 +142,3 @@ export interface CmsProductFacetNavigationComponent extends CmsComponent {
   searchResult?: string;
   minPerFacet?: string;
 }
-
-export interface CmsMerchandisingCarouselComponent extends CmsComponent {
-  title?: string;
-  strategy?: string;
-  numberToDisplay?: string;
-  scroll?: string;
-  container?: string;
-  textColour?: string;
-  backgroundColour?: string;
-}
