+++ projects/storefrontlib/src/cms-components/index.ts
@@ -3,7 +3,6 @@ export * from './cart/index';
 export * from './checkout/index';
 export * from './cms-lib.module';
 export * from './content/index';
-export * from './merchandising/index';
 export * from './misc/index';
 export * from './myaccount/index';
 export * from './navigation/index';
